==> 04_Model_Preprocessing.ipynb <==
# Oddstradamus
### Good odds and where to find them

### Model Preprocessing

In order to gain an overview of which features are ultimately used for model building after the EDA, each feature is first listed and described. Then the split of the training and test data is carried out for all 6 data sets.

#### Model Features

| Feature | Description |
| - | - |
| `AvgHome` | Average Odd Home Win |
| `AvgDraw` | Average Odd Draw |
| `AvgAway` | Average Odd Away Win |
| `MaxHome` | Maximum Odd Home Win |
| `MaxDraw` | Maximum Odd Draw |
| `MaxAway` | Maximum Odd Away Win |
| `GoalsHome_before` | Number of goals home team scored at the start of the match (in home matches) |
| `GoalsAway_before` | Number of goals away team scored at the start of the match (in away matches) |
| `Home_Goals_Conceded_before` | Number of goals home team conceded at the start of the match (in Home matches) |
| `Away_Goals_Conceded_before` | Number of goals home team conceded at the start of the match (in Away matches) |
| `TotalGoals_Home_before` | Total number of goals in home matches |
| `TotalGoals_Away_before` | Total number of goals in away matches |
| `ShotsHome_before` | Number of shots home team scored at the start of the match |
| `ShotsAway_before` | Number of shots away team scored at the start of the match |
| `Home_Shots_Conceded_before` | Number of shots home team conceded at the start of the match |
| `Away_Shots_Conceded_before` | Number of shots away team conceded at the start of the match |
| `TotalShots_Home_before` | Total number of shots in home matches |
| `TotalShots_Away_before` | Total number of shots in away matches |
| `SOTHome_before` | Number of shots on target home team scored at the start of the match |
| `SOTAway_before` | Number of shots on target away team scored at the start of the match |
| `Home_SOT_Conceded_before` | Number of shots home team conceded at the start of the match |
| `Away_SOT_Conceded_before` | Number of shots away team conceded at the start of the match |
| `TotalSOT_Home_before` | Total number of shots on target in home matches |
| `TotalSOT_Away_before` | Total number of shots on target in home matches |
| `CornersHome_before` | Number of corners home team scored at the start of the match |
| `CornersAway_before` | Number of corners away team scored at the start of the match |
| `Home_Corners_Conceded_before` | Number of corners home team conceded at the start of the match |
| `Away_Corners_Conceded_before` | Number of corners away team conceded at the start of the match |
| `TotalCorners_Home_before` | Total number of corners in home matches |
| `TotalCorners_Away_before` | Total number of corners in away matches |
| `home_wins_season` | Number of home wins in the season (Home team) |
| `home_draws_season` | Number of home draws in the season (Home team) |
| `home_loses_season` | Number of home loses in the season (Home team) |
| `away_wins_season` | Number of home wins in the season (Away team) |
| `away_draws_season` | Number of home draws in the season (Away team) |
| `away_loses_season` | Number of home losess in the season (Away team) |
| `home_points` | Number of points scored in home matches (Home team) |
| `away_points` | Number of points scored in home matches (Away team) |
| `homewin_avg_odd` | Average win odd of the home team |
| `homedraw_avg_odd` | Average draw odd of the home team |
| `homelose_avg_odd` | Average lose odd of the home team |
| `awaywin_avg_odd` | Average win odd of the away team |
| `awaydraw_avg_odd` | Average draw odd of the away team |
| `awaylose_avg_odd` | Average lose odd of the away team |
| `homewin_avg_odd_season` | Average win odd of the home team per season |
| `homedraw_avg_odd_season` | Average draw odd of the home team per season |
| `homelose_avg_odd_season` | Average lose odd of the home team per season |
| `awaywin_avg_odd_season` | Average win odd of the away team per season |
| `awaydraw_avg_odd_season`| Average draw odd of the away team per season |
| `awaylose_avg_odd_season` | Average lose odd of the away team per season |
| `home_prev_win_all` | Average win odd of the home team if the home win occurs |
| `home_prev_draw_all` | Average draw odd of the home team if the draw occurs |
| `home_prev_lose_all` | Average lose odd of the home team if the home lose occurs|
| `away_prev_win_all` | Average win odd of the away team if the away win occurs |
| `away_prev_draw_all` | Average draw odd of the away team if the draw occurs |
| `away_prev_lose_all` | Average lose odd of the away team if the away lose occurs |
| `home_prev_win_season` | Average win odd of the home team if the home win occurs per season |
| `home_prev_draw_season` | Average draw odd of the home team if the draw occurs per season  |
| `home_prev_lose_season` | Average lose odd of the home team if the home lose occurs per season |
| `away_prev_win_season` | Average win odd of the away team if the away win occurs per season |
| `away_prev_draw_season` | Average draw odd of the away team if the draw occurs per season |
| `away_prev_lose_season` | Average lose odd of the away team if the away lose occurs per season |
| `team_diff_goals_home_before` | Team-specific goal difference in home matches |
| `team_diff_shots_home_before` | Team-specific shot difference in home matches |
| `team_diff_SOT_home_before` | Team-specific shots on target difference in home matches |
| `team_diff_corners_home_before` | Team-specific corner difference in home matches |
| `team_diff_goals_away_before` | Team-specific goal difference in away matches |
| `team_diff_shots_away_before` | Team-specific shot difference in away matches |
| `team_diff_SOT_away_before` | Team-specific shots on target difference in away matches |
| `team_diff_corners_away_before` | Team-specific corner difference in away matches |
| `cross_diff_goals_before` | Match Rating: Goal difference of home team in home matches and away team in away matches |
| `cross_diff_shots_before` | Match Rating: Shot difference of home team in home matches and away team in away matches |
| `cross_diff_SOT_before` | Match Rating: Shots on target difference of home team in home matches and away team in away matches |
| `cross_diff_corners_before`| Match Rating: Corners difference of home team in home matches and away team in away matches |
| `cross_goal_home` | Match Rating: Difference between goals scored by home team in home matches and goals conceded by away team in away matches |
| `cross_shots_home` | Match Rating: Difference between shots scored by home team in home matches and shots conceded by away team in away matches |
| `cross_SOT_home` | Difference between shots on target scored by home team in home matches and shots on target conceded by away team in away matches |
| `cross_corners_home` | Difference between corners scored by home team in home matches and corners conceded by away team in away matches |
| `cross_goal_away` | Match Rating: Difference between goals scored by away team in away matches and goals conceded by home team in home matches |
| `cross_shots_away` | Match Rating: Difference between shots scored by away team in away matches and shots conceded by home team in home matches |
| `cross_SOT_away` | Match Rating: Difference between shots on target scored by away team in away matches and shots on target conceded by home team in home matches |
| `cross_corners_away` | Match Rating: Difference between corners scored by away team in away matches and corners conceded by home team in home matches |
| `cross_diff_goals_Conceded_before` | Match Rating: Difference between goals conceded by home team in home matches and goals conceded by away team in away matches |
| `cross_diff_shots_Conceded_before` | Match Rating: Difference between shots conceded by home team in home matches and shots conceded by away team in away matches |
| `cross_diff_SOT_Conceded_before` | Match Rating: Difference between shots on target conceded by home team in home matches and shots on target conceded by away team in away matches |
| `cross_diff_corners_Conceded_before` | Match Rating: Difference between corners conceded by home team in home matches and corners conceded by away team in away matches |
| `two_way_h_odd` | Two-way-result (1 = Home win, X2 = draw or away win) TARGET IN HOME DATA SETS |
| `two_way_a_odd` | Two-way-result (2 = Away win, 1X = home win or draw) TARGET IN AWAY DATA SETS |
| `dc_away` | odd for X2-result (only in home data sets) |
| `dc_home` | odd for 1X-result (only in away data sets) |
| `diff_exp_val_home` | Difference between home win odd and average home win odd if home win occurs |
| `diff_exp_val_away` | Difference between away win odd and average home win odd if away win occurs |
| `cross_diff_points_before`| Match Rating: Difference between home points of the home team and away points of the away team |
| `Div_` | Total 21 dummie-variables. One for each Division |
| `Season_` | Total 7 dummie-variables. One for each Season |

#### Train-Test-Split

# import packages
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split

##### 50%-Dataframe

# loading the modified dataframes
df_h = pd.read_csv('Data/model_home.csv')
df_a = pd.read_csv('Data/model_away.csv');

# assign features and target
X_h = df_h.drop('two_way_h_odd', axis=1)
y_h = df_h.two_way_h_odd

# split the dataset
X_train_h, X_test_h, y_train_h, y_test_h = train_test_split(X_h, y_h, test_size=0.3, random_state=42, stratify = y_h)

# save the training and test sets
X_train_h.to_csv('Data/model_home_csv/X_train_home_50.csv', index = False)
X_test_h.to_csv('Data/model_home_csv/X_test_home_50.csv', index = False)
y_train_h.to_csv('Data/model_home_csv/y_train_home_50.csv', index = False)
y_test_h.to_csv('Data/model_home_csv/y_test_home_50.csv', index = False)

# assign features and target
X_a = df_a.drop('two_way_a_odd', axis=1)
y_a = df_a.two_way_a_odd

# split the dataset
X_train_a, X_test_a, y_train_a, y_test_a = train_test_split(X_a, y_a, test_size=0.3, random_state=36, stratify = y_a)

# save the training and test sets
X_train_a.to_csv('Data/model_away_csv/X_train_away_50.csv', index = False)
X_test_a.to_csv('Data/model_away_csv/X_test_away_50.csv', index = False)
y_train_a.to_csv('Data/model_away_csv/y_train_away_50.csv', index = False)
y_test_a.to_csv('Data/model_away_csv/y_test_away_50.csv', index = False)

##### Favorite-Dataframe

# loading the modified dataframes
df_h_fav = pd.read_csv('Data/model_home_fav.csv')
df_a_fav = pd.read_csv('Data/model_away_fav.csv')

# assign features and target
X_h_fav = df_h_fav.drop('two_way_h_odd', axis=1)
y_h_fav = df_h_fav.two_way_h_odd

# split the dataset
X_train_h_fav, X_test_h_fav, y_train_h_fav, y_test_h_fav = train_test_split(X_h_fav, y_h_fav, test_size=0.3, random_state=42, stratify = y_h_fav)

# save the training and test sets
X_train_h_fav.to_csv('Data/model_home_csv/X_train_home_fav.csv', index = False)
X_test_h_fav.to_csv('Data/model_home_csv/X_test_home_fav.csv', index = False)
y_train_h_fav.to_csv('Data/model_home_csv/y_train_home_fav.csv', index = False)
y_test_h_fav.to_csv('Data/model_home_csv/y_test_home_fav.csv', index = False)

# assign features and target
X_a_fav = df_a_fav.drop('two_way_a_odd', axis=1)
y_a_fav = df_a_fav.two_way_a_odd

# split the dataset
X_train_a_fav, X_test_a_fav, y_train_a_fav, y_test_a_fav = train_test_split(X_a_fav, y_a_fav, test_size=0.3, random_state=42, stratify = y_a_fav)

# save the training and test sets
X_train_a_fav.to_csv('Data/model_away_csv/X_train_away_fav.csv', index = False)
X_test_a_fav.to_csv('Data/model_away_csv/X_test_away_fav.csv', index = False)
y_train_a_fav.to_csv('Data/model_away_csv/y_train_away_fav.csv', index = False)
y_test_a_fav.to_csv('Data/model_away_csv/y_test_away_fav.csv', index = False)

##### Full-Dataframe

# loading the modified dataframe
df_full = pd.read_csv('Data/model_full.csv')

# assign features and target
X_h_full = df_full.drop(['two_way_h_odd','two_way_a_odd'], axis=1)
y_h_full = df_full.two_way_h_odd

# split the dataset
X_train_h_full, X_test_h_full, y_train_h_full, y_test_h_full = train_test_split(X_h_full, y_h_full, test_size=0.3, random_state=42, stratify = y_h_full)

# save the training and test sets
X_train_h_full.to_csv('Data/model_home_csv/X_train_home_full.csv', index = False)
X_test_h_full.to_csv('Data/model_home_csv/X_test_home_full.csv', index = False)
y_train_h_full.to_csv('Data/model_home_csv/y_train_home_full.csv', index = False)
y_test_h_full.to_csv('Data/model_home_csv/y_test_home_full.csv', index = False)

# assign features and target
X_a_full = df_full.drop(['two_way_h_odd','two_way_a_odd'], axis=1)
y_a_full = df_full.two_way_a_odd

# split the dataset
X_train_a_full, X_test_a_full, y_train_a_full, y_test_a_full = train_test_split(X_a_full, y_a_full, test_size=0.3, random_state=42, stratify = y_a_full)

# save the training and test sets
X_train_a_full.to_csv('Data/model_away_csv/X_train_away_full.csv', index = False)
X_test_a_full.to_csv('Data/model_away_csv/X_test_away_full.csv', index = False)
y_train_a_full.to_csv('Data/model_away_csv/y_train_away_full.csv', index = False)
y_test_a_full.to_csv('Data/model_away_csv/y_test_away_full.csv', index = False)
